first working version for mandatory part. Waiting on clarification for last line what expected outcome is

+++ get_next_line/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: mservais <mservais@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/07 16:01:50 by mservais          #+#    #+#             */
-/*   Updated: 2021/07/12 10:28:37 by mservais         ###   ########.fr       */
+/*   Updated: 2021/07/12 13:55:20 by mservais         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ char	*ft_strdup_new_line(char const *s);
 char	*ft_substr(char *s, unsigned int start, size_t len);
 char	*ft_error_check(int fd);
 int		ft_end_of_line(char const *str);
-int		ft_find_newline(char const *str);
+size_t	ft_find_newline(char const *str);
 char	*ft_return(char **line_in_memory, char **buffer);
 char	*get_next_line(int fd);
